backend/requirements.txt
Django==2.2.6
psycopg2-binary==2.8.4
djangorestframework==3.10.3
requests==2.22.0
pylint-django==2.0.13
PyJWT==1.7.1
coverage==5.0.3
django-cors-headers==3.2.1
python-memcached==1.59
statsmodels==0.11.1
sklearn
pandas